pixyz_worker/requirements.txt
redis
celery
celery[redis]
celery[tblib]
tblib
gevent
eventlet
flower
azure-storage-blob
python-multipart
pytest
coverage
pytest-md-report
python-dotenv
requests-toolbelt